requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
langchain
langchain-groq
langchain-huggingface
groq
python-dotenv
faiss-cpu
